!823 适配hcom修改

+++ src/hybm/csrc/under_api/hcom_service_c_define.h
@@ -290,7 +290,7 @@ typedef struct {
 } Channel_OneSideRequest;
 
 typedef struct {
-    uint64_t intervalTimeMs;
+    uint16_t intervalTimeMs;
     uint64_t thresholdByte;
     Channel_FlowCtrlLevel flowCtrlLevel;
 } Channel_FlowCtrlOptions;
